+++ packages/PresetElementPlus/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @form-renderer/preset-element-plus
 
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- 修复预设样式问题
+
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- 修复预设样式问题
+
+## 1.0.0-alpha.2
+
+### Patch Changes
+
+- 修复预设样式问题
+
+## 1.0.0-alpha.1
+
+### Patch Changes
+
+- 修复预设样式问题
+
 ## 1.0.0-alpha.0
 
 ### Major Changes