+++ packages/StarterElementPlus/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @form-renderer/starter-element-plus
 
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- 修复预设样式问题
+- Updated dependencies
+  - @form-renderer/preset-element-plus@1.0.0-alpha.4
+
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- 修复预设样式问题
+- Updated dependencies
+  - @form-renderer/preset-element-plus@1.0.0-alpha.3
+
+## 1.0.0-alpha.2
+
+### Patch Changes
+
+- 修复预设样式问题
+- Updated dependencies
+  - @form-renderer/preset-element-plus@1.0.0-alpha.2
+
+## 1.0.0-alpha.1
+
+### Patch Changes
+
+- 修复预设样式问题
+
+### Patch Changes
+
+- Updated dependencies
+  - @form-renderer/preset-element-plus@1.0.0-alpha.1
+
 ## 1.0.0-alpha.0
 
 ### Major Changes